Extract and create InstancesComponent
