+++ src/app/modules/data-set/data-set.component.ts
@@ -11,7 +11,7 @@ import { ExportDraftDataSetDialogComponent } from "./dialogs/export-draft-data-s
 import { UpdateDataSetDialogComponent } from "./dialogs/update-data-set-dialog/update-data-set-dialog.component";
 import { DataSet } from "./model/data-set/data-set.model";
 import { DataSetService } from "./services/data-set.service";
-import { AnnotationNotificationService } from "./services/shared/annotation-notification.service";
+import { NotificationService } from "./services/shared/notification.service";
 import { LocalStorageService } from "./services/shared/local-storage.service";
 
 
@@ -34,7 +34,7 @@ export class DataSetComponent implements OnInit {
 
     constructor(private dialog: MatDialog, private toastr: ToastrService, 
         private datasetService: DataSetService, private router: Router,
-        private storageService: LocalStorageService, private annotationNotificationService: AnnotationNotificationService) {}
+        private storageService: LocalStorageService, private annotationNotificationService: NotificationService) {}
 
     public async ngOnInit(): Promise<void> {
         if (!this.storageService.getLoggedInAnnotator()) this.router.navigate(['/login']);
