+++ src/app/modules/data-set/services/shared/notification.service.ts
@@ -1,6 +1,5 @@
 import { Injectable, EventEmitter } from '@angular/core';
 import { Annotation } from '../../model/annotation/annotation.model';
-import { DataSetProject } from '../../model/data-set-project/data-set-project.model';
 import { DataSet } from '../../model/data-set/data-set.model';
 import { Instance } from '../../model/instance/instance.model';
 
@@ -8,13 +7,16 @@ import { Instance } from '../../model/instance/instance.model';
 @Injectable({
   providedIn: 'root'
 })
-export class AnnotationNotificationService {
+
+export class NotificationService {
 
   public newAnnotation: EventEmitter<Annotation> = new EventEmitter<Annotation>();
   public changedAnnotation: EventEmitter<Annotation> = new EventEmitter<Annotation>();
   public datasetChosen: EventEmitter<DataSet> = new EventEmitter<DataSet>();
   public projectChosen: EventEmitter<any> = new EventEmitter<any>();
   public instanceChosen: EventEmitter<Instance> = new EventEmitter<Instance>();
+  public nextInstance: EventEmitter<void> = new EventEmitter<void>();
+  public previousInstance: EventEmitter<void> = new EventEmitter<void>();
 
   constructor() { }
